pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-rag-prototype"
version = "0.1.0"
description = "NIM-ready edge RAG prototype for RTX 4060"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "chromadb>=0.5.5",
  "sentence-transformers>=3.0.0",
  "pypdf>=4.3.1",
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
edge-rag = "edge_rag.cli:main"
